+++ src/main/java/com/example/ISAprojekat/Model/DTO/FastReservationDTO.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class FastReservationDTO {
-    private Long boatId;
+    private Integer boatId;
     private Date startDate;
     private int duration;
     private int capacity;
