+++ src/main/java/com/example/ISAprojekat/Repository/OcenaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OcenaRepository extends JpaRepository<Ocena,Long> {
+public interface OcenaRepository extends JpaRepository<Ocena,Integer> {
 }
