+++ src/main/java/com/example/ISAprojekat/Service/BoatService.java
@@ -9,9 +9,9 @@ public interface BoatService {
 
 
     List<Boat> findAll();
-    Boat getOne(Long id);
+    Boat getOne(Integer id);
     Boat create(Boat boat) throws Exception;
-    void delete(Long id);
+    void delete(Integer id);
     Boat update(Boat boat);
