+++ src/main/java/com/example/ISAprojekat/Service/OcenaService.java
@@ -5,8 +5,8 @@ import com.example.ISAprojekat.Model.Ocena;
 import java.util.List;
 
 public interface OcenaService {
-    Ocena findOne(Long id);
+    Ocena findOne(Integer id);
     List<Ocena> findAll();
-    float srednjaBrod(List<Ocena> ocenas, Long boat_id);
-    float srednjaVikendica(List<Ocena> ocenas, Long cott_id);
+    float srednjaBrod(List<Ocena> ocenas, Integer boat_id);
+    float srednjaVikendica(List<Ocena> ocenas, Integer cott_id);
 }
